requirements.txt
asgiref==3.3.4
audioread==2.1.9
Django==3.2
gunicorn==20.1.0
mysql-connector-python==8.0.23
protobuf==3.15.8
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
